hospital/requirements.txt
mysql-connector-python
pillow
tkcalendar
openpyxl
fpdf2
